Lixeira finalizada.

+++ app/helpers/application_helper.rb
@@ -13,16 +13,19 @@ module ApplicationHelper
   # Campo com imagens V ou X para habilitar/desabilitar e degradê se não tiver permissão para alteração.
   def toggle_field(obj, field, action='toggle_field', options = {})
     ''.tap do |menu|
-      if check_permission(controller.class, "#{action}")
+      if check_permission(controller.class, "#{action}") and (!obj.respond_to?(:deleted) or (obj.respond_to?(:deleted) and !obj.try(:deleted)))
+        url_options = {:action => "#{action}", :id => obj.id, :field => "#{field}"}
+        url_options[:controller] = options.delete(:controller) if options[:controller]
+        url_options[:menu_id] = options.delete(:menu_id) if options[:menu_id]
         if obj[field.to_s] == 0 or not obj[field.to_s]
           menu << link_to( image_tag("false.png", :alt => t("disable")),
-            {:action => "#{action}", :id => obj.id, :field => "#{field}"},
+            url_options,
             options.merge({method: :put, :title => t("activate")}))
           menu << " #{t('unpublished')}" if options[:show_label]
         else
           menu << link_to(image_tag("true.png", :alt => t("enable")),
-                          {:action => "#{action}", :id=> obj.id, :field => "#{field}"},
-                          options.merge({method: :put, :title => t("deactivate")}))
+            url_options,
+            options.merge({method: :put, :title => t("deactivate")}))
           menu << " #{t('published')}" if options[:show_label]
         end
       else
@@ -37,6 +40,9 @@ module ApplicationHelper
     end
   end
 
+  def status_icon(obj, field = :publish)
+    image_tag("#{obj.send(field)}.png")
+  end
   # Define os menus
   # Parâmetros: Lista de menu (sons, view_ctrl=false)
   # html_class: "dropdown" ou "expanded"
@@ -58,6 +64,7 @@ module ApplicationHelper
 
   # Método recursivo para gerar submenus e os controles
   def print_menu_entry(sons, entry, view_ctrl, indent=0)
+    return "" if !view_ctrl && !entry.publish
     indent_space = " " * indent
     submenu = sons[entry.id].present?
     current_page = (@page && @page == entry.target) || request.path == entry.url
@@ -71,6 +78,7 @@ module ApplicationHelper
       #if (entry.menu.try(:page_id).nil? and entry.menu.try(:link).empty?)
         #menus << "#{entry.menu.try(:title)}"
       #else
+        menus << (entry.deleted ? status_icon(entry) : toggle_field(entry, "publish", :toggle_field, {controller: "sites/admin/menus/menu_items", menu_id: entry.menu_id}) ) if view_ctrl
         menus << link_to(entry.title, entry.target_id.to_i > 0 ? main_app.site_page_path(entry.target_id) : entry.url, :alt => entry.title,:title => entry.description, :target => entry.new_tab ? "_blank":"")
       #end
 
@@ -85,13 +93,15 @@ module ApplicationHelper
             end
           end
         end
-        #menus << " [ id:#{entry.id} pos:#{entry.position} ]" # Para debug
-        menus << ( (entry and entry.target) ? " [ #{entry.target.try(:title)} ] " : " [ #{entry.url if not entry.url.blank?} ] " )
-        menus << link_to(icon('edit', text: ''), edit_site_admin_menu_menu_item_path(entry.menu_id, entry.id), :title => t("edit"))
-        menus << indent_space + link_to(icon('plus', text: ''), new_site_admin_menu_menu_item_path(entry.menu_id, :parent_id => entry.id), :title => t("add_sub_menu"))
-        menus << indent_space + link_to(icon('fire', text: ''), site_admin_menu_menu_item_path(entry.menu_id, entry.id), :method=>:delete, :data => {:confirm => t('are_you_sure')}, :title => t("destroy"))
-        menus << indent_space + link_to(icon('trash', text: ''), remove_site_admin_menu_menu_item_path(entry.menu_id, entry.id), :title => t("remove"), :data => {:confirm => t("are_you_sure")}) 
-        menus << indent_space + link_to(icon('move', text: ''),"#", :class => 'handle', :title => t("move"))
+        if entry.deleted
+        else
+          #menus << " [ id:#{entry.id} pos:#{entry.position} ]" # Para debug
+          menus << ( (entry and entry.target) ? " [ #{entry.target.try(:title)} ] " : " [ #{entry.url if not entry.url.blank?} ] " )
+          menus << link_to(icon('edit', text: ''), edit_site_admin_menu_menu_item_path(entry.menu_id, entry.id), :title => t("edit"))
+          menus << indent_space + link_to(icon('plus', text: ''), new_site_admin_menu_menu_item_path(entry.menu_id, :parent_id => entry.id), :title => t("add_sub_menu"))
+          menus << indent_space + link_to(icon('fire', text: ''), site_admin_menu_menu_item_path(entry.menu_id, entry.id), :method=>:delete, :data => {:confirm => t('are_you_sure')}, :title => t("destroy"))
+          menus << indent_space + link_to(icon('move', text: ''),"#", :class => 'handle', :title => t("move"))
+        end
       end
       menus << "\n" + indent_space + (view_ctrl ? "</div><menu>":"<menu class='submenu'>") if submenu
       if sons[entry.id].class.to_s == "Array"
